Figured out better take implementation
Looked though the Kotlin source and realized that I could just throw an
exception to cancel the collect. This lead to an easier take because
then other Currents don't need to know that they can be canceled. I will
just cancel them and catch it when it comes back down to me.