Fixed all the detekt warnings
I figured out that I didn't really need the class declarations. I can
just use `current {` to build them. It does cause an extra allocation of
the builder, but it seems rather insignificant. That fixed a lot of the
filename mismatch problem. Then a lot of files were missing a newline at
the end. Finally I had a swallowed exception which was on purpose so I
suppressed it.