Added a buffer with skip, concat, and merge.
During Tech Talk Tuesday, we had a question about prepopulating a buffer
which lead to having to implement the "skip" variant of buffer. The end
solution to the problem was to just concat a starting value to the
current so that you got a value the very first time.

That is also why we implemented a merge and a concat. They only differ
by one running sequentially and one in parallel.

Finally implemented a little helper collectIndex, and withEach.
collectIndex just provides you the index of the value your collecting.
withEach is just a forEach only the value is passed though "with" to
make it the receiver of the function.